+++ lib/layout/news_app/news_layout.dart
@@ -28,7 +28,7 @@ class NewsLayout extends StatelessWidget {
                   icon: Icon(Icons.search)),
               IconButton(
                   onPressed: () {
-                    AppCubit.get(context).changeAppMode();
+                    NewsCubit.get(context).changeAppMode();
                   },
                   icon: Icon(Icons.brightness_4_outlined)),
             ],
