+++ lib/main.dart
@@ -1,11 +1,7 @@
 import 'package:flutter/material.dart';
-import 'package:flutter/services.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
-import 'package:hexcolor/hexcolor.dart';
 import 'package:news_app/layout/news_app/cubit/cubit.dart';
-import 'package:news_app/layout/todo_app/todo.dart';
-import 'package:news_app/shared/cubit/cubit.dart';
-import 'package:news_app/shared/cubit/states.dart';
+import 'package:news_app/layout/news_app/cubit/states.dart';
 import 'package:news_app/shared/network/local/cache_helper.dart';
 import 'package:news_app/shared/network/remote/dio_helper.dart';
 import 'package:news_app/styles/bloc_observ.dart';
@@ -29,33 +25,21 @@ class MyApp extends StatelessWidget {
   // This widget is the root of your application.
   @override
   Widget build(BuildContext context) {
-    return MultiBlocProvider(
-      providers: [
-        BlocProvider(
-            create: (BuildContext context) => NewsCubit()
-              ..getBusiness()
-              ..getSports()
-              ..getScience()),
-        BlocProvider(
-          create: (BuildContext context) => AppCubit()
-            ..changeAppMode(
-              fromShared: isDark,
-            ),
-        ),
-      ],
-      child: BlocConsumer<AppCubit, AppState>(
-        listener: (context, state) {},
-        builder: (context, state) {
-          return MaterialApp(
-            debugShowCheckedModeBanner: false,
-            theme: lightTheme,
-            darkTheme: darkTheme,
-            themeMode:
-                AppCubit.get(context).isDark ? ThemeMode.dark : ThemeMode.light,
-            home: NewsLayout(),
-          );
-        },
-      ),
-    );
+    return BlocProvider(
+  create: (context) => NewsCubit()..getBusiness()..getScience()..getSports(),
+  child: BlocConsumer<NewsCubit, NewsState>(
+      listener: (context, state) {},
+      builder: (context, state) {
+        return MaterialApp(
+          debugShowCheckedModeBanner: false,
+          theme: lightTheme,
+          darkTheme: darkTheme,
+          themeMode:
+              NewsCubit.get(context).isDark ? ThemeMode.dark : ThemeMode.light,
+          home: NewsLayout(),
+        );
+      },
+    ),
+);
   }
 }
