+++ lib/modules/news_app/business/business_screen.dart
@@ -6,6 +6,8 @@ import 'package:news_app/layout/news_app/cubit/cubit.dart';
 import 'package:news_app/layout/news_app/cubit/states.dart';
 
 class BusinessScreen extends StatelessWidget {
+  const BusinessScreen({super.key});
+
   @override
   Widget build(BuildContext context) {
     return BlocConsumer<NewsCubit, NewsState>(
