--- lib/modules/todo_app/done_tasks/doneTask.dart
@@ -1,20 +1,0 @@
-import 'package:flutter/material.dart';
-import 'package:flutter_bloc/flutter_bloc.dart';
-import 'package:news_app/component/components.dart';
-import 'package:news_app/shared/cubit/cubit.dart';
-import 'package:news_app/shared/cubit/states.dart';
-
-class DoneTasksScreen extends StatelessWidget {
-  const DoneTasksScreen({Key? key}) : super(key: key);
-
-  @override
-  Widget build(BuildContext context) {
-    return BlocConsumer<AppCubit, AppState>(
-      listener: (context, state) {},
-      builder: (context, state) {
-        var tasks = AppCubit.get(context).doneTasks;
-        return tasksBuilder(tasks: tasks);
-      },
-    );
-  }
-}
