--- lib/modules/todo_app/new_tasks/newTask.dart
@@ -1,20 +1,0 @@
-
-import 'package:flutter/material.dart';
-import 'package:flutter_bloc/flutter_bloc.dart';
-import 'package:news_app/component/components.dart';
-import 'package:news_app/component/constant.dart';
-import 'package:news_app/shared/cubit/cubit.dart';
-import 'package:news_app/shared/cubit/states.dart';
-
-class NewTasksScreen extends StatelessWidget {
-  @override
-  Widget build(BuildContext context) {
-    return BlocConsumer<AppCubit, AppState>(
-      listener: (context, state) {},
-      builder: (context, state) {
-        var tasks = AppCubit.get(context).newTasks;
-        return tasksBuilder(tasks: tasks);
-      },
-    );
-  }
-}
