+++ lib/shared/cubit/states.dart
@@ -1,21 +1,21 @@
-abstract class AppState{}
-
-class AppInitialState extends AppState{}
-
-class AppChangeBottomNabBarState extends AppState{}
-
-class AppCreatDatabaseState extends AppState{}
-
-class AppGetDatabaseLoadingState extends AppState{}
-
-class AppGetDatabaseState extends AppState{}
-
-class AppInsertDatabaseState extends AppState{}
-
-class AppUpdateDatabaseState extends AppState{}
-
-class AppDeleteDatabaseState extends AppState{}
-
-class AppChangeSheetState extends AppState{}
-
-class AppChangeModeState extends AppState{}
+// abstract class AppState{}
+//
+// class AppInitialState extends AppState{}
+//
+// class AppChangeBottomNabBarState extends AppState{}
+//
+// class AppCreatDatabaseState extends AppState{}
+//
+// class AppGetDatabaseLoadingState extends AppState{}
+//
+// class AppGetDatabaseState extends AppState{}
+//
+// class AppInsertDatabaseState extends AppState{}
+//
+// class AppUpdateDatabaseState extends AppState{}
+//
+// class AppDeleteDatabaseState extends AppState{}
+//
+// class AppChangeSheetState extends AppState{}
+//
+// class AppChangeModeState extends AppState{}
